Dynamic range

getDynamicRange Calculate the dynamic range of the RNAseq experiment

getDynamicRange(dm)

Arguments

dm

The duplication matrix calculated by analyzeDuprates

Value

A list with 2 elements, containing the dynamic range counting all reads and the dynamic range after removing duplicates.

Details

This function calculates the dynamic range of the RNAseq eperiment

Examples

# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)
#> The following objects are masked from dupRadar_examples (pos = 3):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 4):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 5):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 6):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 7):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 8):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 9):
#> 
#>     dm, dm.bad
#> The following objects are masked from dupRadar_examples (pos = 10):
#> 
#>     dm, dm.bad

# calculate the dynamic range
getDynamicRange(dm)
#> $dynrange.all
#> [1] 31153.04
#> 
#> $dynrange.duprm
#> [1] 3343.27
#>